feat: Implement base Gateway schema/handlers
There's now a basic IDENTIFY schema and handler. The Gateway can now
decode an opcode, deserialize the data, and foward it to the correct
handler, and return a close frame. Of which, all of Discord's custom
close frame codes/messages are fully added.

We once again switch another library, this time we're using warp over
tokio-tungstenite as I wasn't able to figure out how to send custom
close frames. Most of the basic code for the Gateway is there now, it's
mostly just schema and the actual opcode implementations that are
required. Though no state is being handled, we might not need state
though? Also some of the code isn't as clean as I would like, looking at
handle_op/mod.rs especially, I would like a better way to get data from
the Gateway message.