Retry at gateway compression and run cargo fmt
Gateway compression is still broken BUT we do
now have fragmentation as we're now interacting
directly with tungstenite types.

Some context for broken compression:
Only the first compressed message is actually
decoding at the client, not sure why. Will have
to investigate later as this is taking up too
much of my time.